requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make freeze
#
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.0
gitdb2==2.0.6             # via gitpython
gitpython==3.0.5
idna==2.8                 # via requests
pyyaml==5.2
requests==2.22.0
smmap2==2.0.5             # via gitdb2
tabulate==0.8.6
urllib3==1.25.7           # via requests
